requirements.txt
click==8.0.1
pyotp==2.6.0
requests==2.25.1
python-nubia==0.2b5
